streamlit-app/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
astor==0.8.1
async-timeout==4.0.3
attrs==23.2.0
azure-ai-vision-imageanalysis==1.0.0b2
azure-core==1.30.1
azure-storage-blob==12.20.0
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.1
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.4
dill==0.3.8
distro==1.9.0
duckdb==0.10.2
entrypoints==0.4
etelemetry==0.3.1
exceptiongroup==1.2.1
faiss-cpu==1.8.0
Faker==19.13.0
favicon==0.7.0
filelock==3.13.3
fitz==0.0.1.dev2
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.6
importlib_metadata==7.1.0
importlib_resources==6.4.0
isodate==0.6.1
Jinja2==3.1.3
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-text-splitters==0.0.1
langsmith==0.1.49
looseversion==1.3.0
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
openai==1.23.2
opencv-python-headless==4.9.0.80
orjson==3.10.1
packaging==23.2
pandas==1.5.3
pandasai==2.0.35
pathlib==1.0.1
pdfminer.six==20231228
pillow==10.3.0
prometheus_client==0.20.0
protobuf==4.25.3
prov==2.0.0
pyarrow==16.0.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
pydot==2.0.0
Pygments==2.17.2
pymdown-extensions==10.8
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pyparsing==3.1.2
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==7.0.0
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
st-annotated-text==4.0.1
streamlit==1.33.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.2
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0
tornado==6.4
tqdm==4.66.2
traits==6.3.2
transformers==4.41.0
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
validators==0.28.1
watchdog==4.0.0
yarl==1.9.4
zipp==3.18.1
azure-storage-blob==12.21.0
